refactor cube files
